Support new (but non-MFA) authentication flow.
This is actually pretty gross, as it's got to do some retries and
sends junk to stdout beacuse there's no logger infrastructure in
place, but Tesla doesn't always send the correct HTTP responses, so
it's rather annoying to figure out what went wrong when part of the
workflow fails.

The good news, I guess, is that we only do this ~once per
application, so it can afford to be a bit yuckier.

This partially fixes #1 -- but there's the MFA workflow which is still
yet more complicated and will also need to be supported.